feat(cli): add "collect" command to find all files and generate entry
At the moment the files in `bin/` are excluded from "prettier" but when
using an editor plugin the files will be formatted anyways which leads
to errors with trailing commas.
We should consider switching prettier's comma mode to "es5" or to run
the files in bin/ through babel.
We could, for example, move all files from `bin/` to `src/bin/` and let
babel write the es5 compatible output to `bin/`.